// src/app/models/clientes.model.ts
export interface Clientes{
    id: number;
    razao_nome: string;
    fantasia_sobrenome: string;
    cnpj_cpf: string;
    ie_rg: string;
    logradouro: string;
    cep: string;
    bairro: string;
    cidade: string;
    uf: string;
    ibge: string;
    codigo_pais: string;
    pais: string;
    ativo: string;
    complemento: string;
    numero: string;
    email: string;
    telefone: string;
    celular: string;
    contato: string;
    tipo_empresa: string;
    porte: string;
    data_abertura: Date;
    natureza_juridica: string;
    tipo: string;
}